Fix build trigger for path filters
When using the path filters, GitHub generates a list of changed files
differently for pull request and push triggers:

- Pull request: diff between most recent commit and merge base
- Push: diff between head and base commit

This means our builds were not triggered if the most recent push didn't
contain changes to the corresponding files, even if the overall pull
request did include changes to them.

This changes the logic to use a pull request based trigger. Pushes
itself won't trigger builds now. If this turns out to be an issue, we
can enable both.